+++ app/Services/ExoProcessDoc.php
@@ -2,6 +2,7 @@
 
 namespace App\Services;
 
+use App\Models\SoalConstant;
 use Illuminate\Support\Str;
 use Intervention\Image\Facades\Image;
 use Illuminate\Support\Facades\Storage;
@@ -212,21 +213,26 @@ class ExoProcessDoc
                 foreach($body->childNodes as $table) {
                     if ($table->nodeName == "table") {
                         $element = [
-                            'pertanyaan' => '',
-                            'correct'    => [],
-                            'options'    => [],
-                            'type'       => 0,
+                            'pertanyaan'            => '',
+                            'correct'               => [],
+                            'correct_benar_salah'   => [],
+                            'options'               => [],
+                            'options_menjodohkan'   => [],
+                            'options_mengurutkan'   => [],
+                            'type'                  => 0,
                         ];
                         foreach($table->childNodes as $iterate => $tr) {
                             $td = $tr->childNodes;
                             $key = $td->item(0);
                             $value = $td->item(1);
 
+                            # FIRST ROW ADALAH PERTANYAAN
                             if ($iterate == 0) {
                                 $element['pertanyaan'] = $this->_dom_inner_html($value);
                                 continue;
                             }
 
+                            # CEK KUNCI JAWABAN
                             if ($key->nodeValue == ":::") {
                                 $correct_op = array_filter(explode(',',$value->nodeValue));
                                 $correct_option_position = array();
@@ -237,17 +243,57 @@ class ExoProcessDoc
                                 $element['correct'] = $correct_option_position;
                                 continue;
                             }
+                            # TIPE SOAL MENJODOHKAN
+                            if(is_numeric(trim($key->nodeValue))) {
+                                $element['options_menjodohkan'][trim($key->nodeValue)][] = $this->_dom_inner_html($value);
+                                continue;
+                            }
+                            # TIPE SOAL MENGURUTKAN
+                            if(trim($key->nodeValue) == '|') {
+                                $element['options_mengurutkan'][] = $this->_dom_inner_html($value);
+                                continue;
+                            }
+                            # TIPE SOAL BENAR SALAH
+                            if(trim($key->nodeValue) == 'BENAR' || trim($key->nodeValue) == 'SALAH') {
+                                array_push($element['options'], $this->_dom_inner_html($value));
+
+                                if(trim($key->nodeValue) == 'BENAR') {
+                                    $element['correct_benar_salah'][count($element['options'])-1] = 1;
+                                } else {
+                                    $element['correct_benar_salah'][count($element['options'])-1] = 0;
+                                }
+                                $element['type'] = SoalConstant::TIPE_BENAR_SALAH;
+
+                                continue;
+                            }
+                            if(trim($key->nodeValue) == '?') {
+                                $element['type'] = SoalConstant::TIPE_SETUJU_TIDAK;
+                            }
                             array_push($element['options'], $this->_dom_inner_html($value));
                         }
-                        if (count($element['correct']) > 1) {
-                            $element['type'] = 4;
-                        } else if (count($element['correct']) == 1) {
-                            $element['type'] = 1;
-                        } else if (count($element['correct']) == 0) {
-                            if (count($element['options']) > 0) {
-                                $element['type'] = 6;
-                            } else {
-                                $element['type'] = 2;
+                        if($element['type'] == 0) {
+                            if (count($element['correct']) > 1) {
+                                $element['type'] = SoalConstant::TIPE_PG_KOMPLEK;
+
+                            } else if (count($element['correct']) == 1) {
+                                $element['type'] = SoalConstant::TIPE_PG;
+
+                            } else if (count($element['correct']) == 0) {
+                                if (count($element['options']) > 0) {
+                                    $element['type'] = SoalConstant::TIPE_ISIAN_SINGKAT;
+
+                                } else {
+                                    if(count($element['options_menjodohkan']) > 0) {
+                                        $element['type'] = SoalConstant::TIPE_MENJODOHKAN;
+                                        
+                                    } else if(count($element['options_mengurutkan']) > 0) {
+                                        $element['type'] = SoalConstant::TIPE_MENGURUTKAN;
+
+                                    } else {
+                                        $element['type'] = SoalConstant::TIPE_ESAY;
+
+                                    }
+                                }
                             }
                         }
                         array_push($data, $element);