+++ app/Services/Ujian/MenjodohkanService.php
@@ -6,6 +6,8 @@ use App\Actions\SendResponse;
 use App\Models\SoalConstant;
 use Exception;
 use Illuminate\Http\Response;
+use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 
@@ -37,6 +39,8 @@ class MenjodohkanService implements TipeSoalInterface
             ]);
             if($setting['acak_soal'] == "1") {
                 $menjodohkan = $menjodohkan->inRandomOrder();
+            } else {
+                $menjodohkan = $menjodohkan->orderBy('created_at');
             }
             $menjodohkan = $menjodohkan->take($max_menjodohkan)->get();
 
@@ -68,8 +72,26 @@ class MenjodohkanService implements TipeSoalInterface
     public static function setJawab($request, $jawaban_peserta)
     {
         $jwb_soals = DB::table('jawaban_soals')
-            ->where('soal_id', $jawaban_peserta->soal_id)
-            ->get();
+            ->where('soal_id', $jawaban_peserta->soal_id);
+
+        if(config('exo.enable_cache')) {
+            $cacheKeyConsolidate = "jawaban_soals_3LHVWCOYFM_".$jawaban_peserta->soal_id;
+            if(Cache::has($cacheKeyConsolidate)) {
+                $jwb_soals = Cache::get($cacheKeyConsolidate, new Collection());
+
+            } else {
+                $jwb_soals = $jwb_soals->get();
+                if($cacheKeyConsolidate) {
+                    Cache::put($cacheKeyConsolidate, $jwb_soals, 60);
+
+                }
+
+            }
+        } else {
+            $jwb_soals = $jwb_soals->get();
+        }
+
+
         $menjodohkan_correct = $jwb_soals->map(function($item) {
             $obj = json_decode($item->text_jawaban, true);
             return [$obj['a']['id'], $obj['b']['id']];
@@ -77,7 +99,7 @@ class MenjodohkanService implements TipeSoalInterface
         $count_corect = 0;
         foreach ($request->menjodohkan as $result) {
             foreach ($menjodohkan_correct as $item) {
-                if ($result == $item) {
+                if (($result[0] == $item[0]) && ($result[1] == $item[1])) {
                     $count_corect += 1;
                     break;
                 }
@@ -89,8 +111,8 @@ class MenjodohkanService implements TipeSoalInterface
         }
         try {
             $data_update = [
-                'iscorrect' => $result_menjodohkan_correct,
-                'menjodohkan' => json_encode($request->menjodohkan)
+                'iscorrect'     => $result_menjodohkan_correct,
+                'menjodohkan'   => json_encode($request->menjodohkan)
             ];
             if (!$jawaban_peserta->answered) {
                 $data_update['answered'] = true;