+++ resources/js/pages/siswa/Index.vue
@@ -27,40 +27,43 @@
       :is-full-page="true"
     ><div class="text-xl text-center">Akun anda dibekukan karena: <u>{{ peserta.block_reason }}</u> <br /> Hubungi administrator untuk membuka blok</div></loading>
     
-    <div class="pt-6 pb-24 shadow-sm border-gray-300 px-4 bg-gradient-to-r from-blue-500 to-blue-400 text-white">
-      <div class="flex justify-between flex-col sm:flex-row">
-        <div class="flex items-center space-x-1"
-        v-if="typeof setting.sekolah != 'undefined'"
-        >
-          <div class="h-16 w-16 bg-white py-1 px-1 rounded-md items-center justify-center flex">
-            <img
-            :src="setting.sekolah.logo != ''
-            ? '/storage/'+setting.sekolah.logo
-            : '/img/exo.jpg'"
-            class="h-12 w-12 object-cover"
-            />
-          </div>
-          <div class="flex flex-col">
-            <p class="font-semibold">{{ setting.sekolah.nama != '' ? setting.sekolah.nama : 'ExtraordinaryCBT' }}</p>
-            <p class="text-sm text-gray-100">CBT-Application</p>
-          </div>
-        </div>
-        <div class="flex space-x-2 justify-end">
-          <div class="flex flex-col">
-            <p class="font-semibold text-right">{{ peserta.nama }}</p>
-            <p class="text-sm text-right">{{ peserta.no_ujian }}</p>
-          </div>
-          <button class="h-12 w-12 flex items-center justify-center bg-white text-gray-600 rounded-md hover:shadow-lg"
-          :class="{ isLoading : 'bg-gray-100' }"
-          :disabled="isLoading"
-          @click="logout"
+    <div v-if="show_disable_screen">Layar harus full-screen untuk mengerjakan exo-cbt ini</div>
+    <template v-if="!show_disable_screen">
+      <div class="pt-6 pb-24 shadow-sm border-gray-300 px-4 bg-gradient-to-r from-blue-500 to-blue-400 text-white">
+        <div class="flex justify-between flex-col sm:flex-row">
+          <div class="flex items-center space-x-1"
+          v-if="typeof setting.sekolah != 'undefined'"
           >
-            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="h-7 feather feather-log-out"><path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4"></path><polyline points="16 17 21 12 16 7"></polyline><line x1="21" y1="12" x2="9" y2="12"></line></svg>
-          </button>
+            <div class="h-16 w-16 bg-white py-1 px-1 rounded-md items-center justify-center flex">
+              <img
+              :src="setting.sekolah.logo != ''
+              ? '/storage/'+setting.sekolah.logo
+              : '/img/exo.jpg'"
+              class="h-12 w-12 object-cover"
+              />
+            </div>
+            <div class="flex flex-col">
+              <p class="font-semibold">{{ setting.sekolah.nama != '' ? setting.sekolah.nama : 'ExtraordinaryCBT' }}</p>
+              <p class="text-sm text-gray-100">CBT-Application</p>
+            </div>
+          </div>
+          <div class="flex space-x-2 justify-end">
+            <div class="flex flex-col">
+              <p class="font-semibold text-right">{{ peserta.nama }}</p>
+              <p class="text-sm text-right">{{ peserta.no_ujian }}</p>
+            </div>
+            <button class="h-12 w-12 flex items-center justify-center bg-white text-gray-600 rounded-md hover:shadow-lg"
+            :class="{ isLoading : 'bg-gray-100' }"
+            :disabled="isLoading"
+            @click="logout"
+            >
+              <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="h-7 feather feather-log-out"><path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4"></path><polyline points="16 17 21 12 16 7"></polyline><line x1="21" y1="12" x2="9" y2="12"></line></svg>
+            </button>
+          </div>
         </div>
       </div>
-    </div>
-    <router-view></router-view>
+      <router-view></router-view>
+    </template>
     <div class="fixed bottom-0 left-0 w-full border-t border-gray-300 text-gray-600 py-2 px-4 text-center bg-white">
 	    <span class="text-sm">&copy; 2019 - {{ year }} Extraordinary CBT {{ version }}</span>
     </div>
@@ -87,7 +90,11 @@ export default {
       enable_socket: process.env.MIX_ENABLE_SOCKET,
       version: process.env.MIX_APP_VERSION,
       year: '',
-      isBlockedPeserta: false
+      isBlockedPeserta: false,
+      only_fullscreen: false,
+      show_disable_screen: false,
+      out_screen_timer: 0,
+      out_screen_interval: 0
     }
   },
   computed: {
@@ -104,7 +111,7 @@ export default {
   methods: {
     ...mapActions('siswa_jadwal',['ujianAktif']),
     ...mapActions('siswa_auth',['logoutPeserta']),
-    ...mapActions('siswa_ujian',['getPesertaDataUjian', 'getPesertaUjian', 'getUncompleteUjian', 'leaveCounterUjian']),
+    ...mapActions('siswa_ujian',['getPesertaDataUjian', 'getPesertaUjian', 'getUncompleteUjian', 'leaveCounterUjian', 'blockMePlease']),
     showError(err) {
       showSweetError(this, err)
     },
@@ -138,11 +145,30 @@ export default {
           let net = await this.leaveCounterUjian({
             id: this.detail.id
           })
-          console.log(net.data.status)
           if(net.data.status == '0') {
             this.$store.commit('siswa_user/_block_peserta', net.data.block_reason)
             this.isBlockedPeserta = true
           }
+        }
+      }
+    },
+    async listenFullScreenChange() {
+      if(this.only_fullscreen) {
+        this.out_screen_timer = 0
+        clearTimeout(this.out_screen_interval)
+        
+        if (window.innerHeight == screen.height) {
+            console.log('FULL SCREEN');
+            this.show_disable_screen = false
+        } else {
+            this.out_screen_interval=setTimeout(async () => {
+              await this.blockMePlease({
+                id: this.detail.id,
+                reason: 'Tidak berada pada mode full-screen'
+              })
+            }, 3000)
+            console.log('NORMAL SCREEN');
+            this.show_disable_screen = true
         }
       }
     }
@@ -209,6 +235,14 @@ export default {
   },
   mounted() {
     document.addEventListener("visibilitychange", this.listenEventChange);
+    document.addEventListener('fullscreenchange', this.listenFullScreenChange);
+    window.addEventListener('resize', this.listenFullScreenChange);
+
+    if(this.setting.only_fullscreen == true) {
+        this.only_fullscreen = true
+
+        this.listenFullScreenChange()
+    }
   },
   watch: {
     uncomplete(val) {
@@ -240,10 +274,19 @@ export default {
           }
         }
       }
+    },
+    setting(v) {
+      if(v.only_fullscreen == true) {
+        this.only_fullscreen = true
+
+        this.listenFullScreenChange()
+      }
     }
   },
   beforeDestroy() {
     document.removeEventListener("visibilitychange", this.listenEventChange)
+    document.removeEventListener("fullscreenchange");
+    window.removeEventListener("resize")
   },
   destroyed() {
     this.$store.commit('siswa_user/REMOTE_PESERTA_DETAIL')