+++ changelog.md
@@ -1,4 +1,21 @@
 ### EXO CBT CHANGE LOG
+### 3.17.0-RELEASE
+- fix issue order when import soal docx 2, docx 1 and html
+- fix issue esay correction not show
+- add banksoal detail tipe inputted
+- fix issue soal random order when not set as random
+- fix created_at not inserted when input soal
+- fix issue preview soal menjodohkan in admin
+- add feature import soal menjodohkan by docx format 2 and format 2 html
+- add feature import soal mengurutkan by docx format 2 and format 2 html
+- add feature import soal benar-salah by docx format 2 and format 2 html
+- add feature import soal setuju-tidak by docx format 2 and format 2 html
+- add feature auto-block when student not in full-screen
+- improve performance cache in answer student
+- add feature edit header text for instantion
+- add kop instantion in berita-acara-event
+- add kop instantion in banksoal-preview
+
 ### 3.16.5-RELEASE
 - endpoint to get peserta in block status
 - endpoint unblock single or multiple
@@ -12,6 +29,7 @@
 
 ### 3.16.2-RELEASE
 - Fix issue with leave-counter
+
 ### 3.16.1-RELEASE
 - Fix issue with leave-counter
 - Anti-block feature